requirements.txt
gunicorn
flask
python-dotenv
pillow
numpy
flask_cors
pymongo